+++ src/main/java/com/netflix/conductor/pheedpgexporter/config/CustomConfiguration.java
@@ -1,16 +1,4 @@
-/*
- * Copyright 2023 Netflix, Inc.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package org.mifos.pheedpgexporter.config;
+package com.netflix.conductor.pheedpgexporter.config;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
@@ -18,8 +6,8 @@ import org.springframework.context.annotation.Configuration;
 
 import com.netflix.conductor.core.listener.TaskStatusListener;
 import com.netflix.conductor.core.listener.WorkflowStatusListener;
-import org.mifos.pheedpgexporter.listeners.CustomTaskStatusListener;
-import org.mifos.pheedpgexporter.listeners.CustomWorkflowStatusListener;
+import com.netflix.conductor.pheedpgexporter.listeners.CustomTaskStatusListener;
+import com.netflix.conductor.pheedpgexporter.listeners.CustomWorkflowStatusListener;
 
 @Configuration
 public class CustomConfiguration {
