--- src/main/java/org/mifos/pheedpgexporter/kafkaexporter/PublishWorkflowMessage.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2023 Netflix, Inc.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package org.mifos.pheedpgexporter.kafkaexporter;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.kafka.core.KafkaTemplate;
-
-import com.netflix.conductor.model.WorkflowModel;import org.springframework.stereotype.Component;
-
-@Component
-public class PublishWorkflowMessage {
-
-    private KafkaTemplate<String, WorkflowModel> kafkaTemplate;
-
-    public void sendMessage(WorkflowModel model) {
-        kafkaTemplate.send("conductor", model.getWorkflowId(), model);
-    }
-}
