+++ src/main/resources/application.yaml
@@ -1,9 +1,3 @@
-conductor:
-  workflow-status-listener:
-    type: custom
-  task-status-listener:
-    type: custom
-
 spring:
   kafka:
     bootstrap-servers: localhost:9092
